+++ tests/test_custommeta.py
@@ -11,11 +11,11 @@ import dbbackup.utils
 TEST_VAL = "aabbcc-1122-3344_eu-west"
 
 
-def dummy_loader():
+def dummy_setter():
     return {"CSMT_VAL": TEST_VAL}
 
 
-def broken_loader():
+def broken_setter():
     return ["not", "a", "dict"]
 
 
@@ -23,7 +23,7 @@ class CSTM:
     abc = "123"
 
 
-def anotherbroken_loader():
+def anotherbroken_setter():
     return {"CSMT_VAL": TEST_VAL, "CSTM_OBJ": CSTM()}
 
 
@@ -47,52 +47,52 @@ def broken_validator(metadata):
 
 # Actual tests
 class CustomMetadataTest(TestCase):
-    @override_settings(DBBACKUP_CUSTOM_METADATA_LOADER="tests.test_custommeta.dummy_loader")
-    def test_metadata_loader_valid(self):
-        """Test that setting DBBACKUP_CUSTOM_METADATA_LOADER works."""
+    @override_settings(DBBACKUP_BACKUP_METADATA_SETTER="tests.test_custommeta.dummy_setter")
+    def test_metadata_setter_valid(self):
+        """Test that setting DBBACKUP_BACKUP_METADATA_SETTER works."""
         importlib.reload(dbbackup.settings)
 
-        assert dbbackup.settings.CUSTOM_METADATA_LOADER == "tests.test_custommeta.dummy_loader"
+        assert dbbackup.settings.BACKUP_METADATA_SETTER == "tests.test_custommeta.dummy_setter"
         assert dbbackup.utils.load_custom_metadata() == {"CSMT_VAL": TEST_VAL}
 
-    @override_settings(DBBACKUP_CUSTOM_METADATA_LOADER="non.existent.loader")
-    def test_metadata_loader_invalid_1(self):
-        """Test that various loader missconfigurations do not work - Non-existent module."""
+    @override_settings(DBBACKUP_BACKUP_METADATA_SETTER="non.existent.loader")
+    def test_metadata_setter_invalid_1(self):
+        """Test that various setter missconfigurations do not work - Non-existent module."""
         importlib.reload(dbbackup.settings)
 
         with pytest.raises(ImportError, match="Could not import module 'non.existent': No module named 'non'"):
             dbbackup.utils.load_custom_metadata()
 
-    @override_settings(DBBACKUP_CUSTOM_METADATA_LOADER="tests.test_custommeta.TEST_VAL")
-    def test_metadata_loader_invalid_2(self):
-        """Test that various loader missconfigurations do not work - Non-callable object."""
+    @override_settings(DBBACKUP_BACKUP_METADATA_SETTER="tests.test_custommeta.TEST_VAL")
+    def test_metadata_setter_invalid_2(self):
+        """Test that various setter missconfigurations do not work - Non-callable object."""
         importlib.reload(dbbackup.settings)
 
         with pytest.raises(ValueError, match="The object at 'tests.test_custommeta.TEST_VAL' is not callable."):
             dbbackup.utils.load_custom_metadata()
 
-    @override_settings(DBBACKUP_CUSTOM_METADATA_LOADER="tests.test_custommeta.broken_loader")
-    def test_metadata_loader_invalid_3(self):
-        """Test that various loader missconfigurations do not work - Wrong return type."""
+    @override_settings(DBBACKUP_BACKUP_METADATA_SETTER="tests.test_custommeta.broken_setter")
+    def test_metadata_setter_invalid_3(self):
+        """Test that various setter missconfigurations do not work - Wrong return type."""
         importlib.reload(dbbackup.settings)
 
-        with pytest.raises(ValueError, match="DBBACKUP_CUSTOM_METADATA_LOADER must return a dictionary."):
+        with pytest.raises(ValueError, match="DBBACKUP_BACKUP_METADATA_SETTER must return a dictionary."):
             dbbackup.utils.load_custom_metadata()
 
-    @override_settings(DBBACKUP_CUSTOM_METADATA_LOADER="tests.test_custommeta.anotherbroken_loader")
-    def test_metadata_loader_invalid_4(self):
-        """Test that various loader missconfigurations do not work - Not JSON serializable."""
+    @override_settings(DBBACKUP_BACKUP_METADATA_SETTER="tests.test_custommeta.anotherbroken_setter")
+    def test_metadata_setter_invalid_4(self):
+        """Test that various setter missconfigurations do not work - Not JSON serializable."""
         importlib.reload(dbbackup.settings)
 
         with pytest.raises(ValueError, match="Custom metadata is not JSON serializable"):
             dbbackup.utils.load_custom_metadata()
 
-    @override_settings(DBBACKUP_CUSTOM_METADATA_VALIDATOR="tests.test_custommeta.dummy_validator")
+    @override_settings(DBBACKUP_RESTORE_METADATA_VALIDATOR="tests.test_custommeta.dummy_validator")
     def test_metadata_validator_valid(self):
-        """Test that setting DBBACKUP_CUSTOM_METADATA_VALIDATOR works."""
+        """Test that setting DBBACKUP_RESTORE_METADATA_VALIDATOR works."""
         importlib.reload(dbbackup.settings)
 
-        assert dbbackup.settings.CUSTOM_METADATA_VALIDATOR == "tests.test_custommeta.dummy_validator"
+        assert dbbackup.settings.RESTORE_METADATA_VALIDATOR == "tests.test_custommeta.dummy_validator"
         assert dbbackup.utils.validate_custom_metadata({"CSMT_VAL": TEST_VAL}) is True
 
         # Test that a validator can raise an exception
@@ -102,7 +102,7 @@ class CustomMetadataTest(TestCase):
         # Test that a wrong also raises a ValueError
         assert dbbackup.utils.validate_custom_metadata({"CSMT_VAL": "aabbcc-xyz_eu-central"}) is False
 
-    @override_settings(DBBACKUP_CUSTOM_METADATA_VALIDATOR="non.existent.validator")
+    @override_settings(DBBACKUP_RESTORE_METADATA_VALIDATOR="non.existent.validator")
     def test_metadata_validator_invalid_1(self):
         """Test that various validator missconfigurations do not work - Non-existent module."""
         importlib.reload(dbbackup.settings)
@@ -110,7 +110,7 @@ class CustomMetadataTest(TestCase):
         with pytest.raises(ImportError, match="Could not import module 'non.existent': No module named 'non'"):
             dbbackup.utils.validate_custom_metadata({"CSMT_VAL": TEST_VAL})
 
-    @override_settings(DBBACKUP_CUSTOM_METADATA_VALIDATOR="tests.test_custommeta.broken_validator")
+    @override_settings(DBBACKUP_RESTORE_METADATA_VALIDATOR="tests.test_custommeta.broken_validator")
     def test_metadata_validator_invalid_2(self):
         """Test that various validator missconfigurations do not work - Exception during validation."""
         importlib.reload(dbbackup.settings)